+++ test/form/samples/base64-deshadow/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'base64 deshadowing indices'
+};