+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Do not emit empty chunks when code-splitting or empty files when preserving modules ([#2128](https://github.com/rollup/rollup/pull/2128))
 * Support `import.meta.url` ([#2164](https://github.com/rollup/rollup/pull/2164))
 * Add `esm` format alias ([#2102](https://github.com/rollup/rollup/pull/2102))
+* Use alphanumeric base64 characters when deconflicting variables ([#2188](https://github.com/rollup/rollup/pull/2188))
 * Improve handling of external modules imported as both default and named imports ([#2136](https://github.com/rollup/rollup/pull/2136))
 * Properly deconflict named imports from other chunks ([#2177](https://github.com/rollup/rollup/pull/2177))
 * Fix an issue with namespaces containing reexports ([#2157](https://github.com/rollup/rollup/pull/2157))